HIVE-29405: moved alterTable() logic from HiveAlterHandler
